add multi-network controller manager
This commit is adding controller manager to manage all controllers of
all networks, preparing for the follow-up changes for the ovn-k8s
multi-network Pod support.
