+++ go-controller/pkg/node/obj_retry_test_helper.go
@@ -154,4 +154,4 @@ func (h *nodePortWatcherEventHandler) RecordDeleteEvent(obj interface{}) {}
 
 func (h *nodePortWatcherEventHandler) RecordSuccessEvent(obj interface{}) {}
 
-func (h *nodePortWatcherEventHandler) RecordErrorEvent(obj interface{}, err error) {}
+func (h *nodePortWatcherEventHandler) RecordErrorEvent(obj interface{}, reason string, err error) {}
