+++ go-controller/pkg/ovn/ovn.go
@@ -2,9 +2,9 @@ package ovn
 
 import (
 	"bytes"
-	"context"
 	"encoding/json"
 	"fmt"
+	ref "k8s.io/client-go/tools/reference"
 	"net"
 	"reflect"
 	"sync"
@@ -14,20 +14,13 @@ import (
 	libovsdbclient "github.com/ovn-org/libovsdb/client"
 
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/config"
-	egressqoslisters "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/crd/egressqos/v1/apis/listers/egressqos/v1"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/factory"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/kube"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/metrics"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/nbdb"
-	addressset "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/address_set"
 	egresssvc "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/controller/egress_services"
 	svccontroller "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/controller/services"
-	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/controller/unidling"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/healthcheck"
-	lsm "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/logical_switch_manager"
-	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/subnetallocator"
-	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/retry"
-	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/syncmap"
 	ovntypes "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/types"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/util"
 
@@ -35,17 +28,12 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/selection"
-	ktypes "k8s.io/apimachinery/pkg/types"
 	apierrors "k8s.io/apimachinery/pkg/util/errors"
 	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/client-go/informers"
 	clientset "k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
-	corev1listers "k8s.io/client-go/listers/core/v1"
-	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/tools/record"
-	ref "k8s.io/client-go/tools/reference"
-	"k8s.io/client-go/util/workqueue"
 	"k8s.io/klog/v2"
 )
 
@@ -59,99 +47,12 @@ type ACLLoggingLevels struct {
 	Deny  string `json:"deny,omitempty"`
 }
 
-// Controller structure is the object which holds the controls for starting
-// and reacting upon the watched resources (e.g. pods, endpoints)
-type Controller struct {
+// BaseNetworkController structure is place holder for all fields shared among controllers.
+type BaseNetworkController struct {
 	client       clientset.Interface
 	kube         kube.Interface
 	watchFactory *factory.WatchFactory
-	stopChan     <-chan struct{}
-
-	// FIXME DUAL-STACK -  Make IP Allocators more dual-stack friendly
-	masterSubnetAllocator        *subnetallocator.HostSubnetAllocator
-	hybridOverlaySubnetAllocator *subnetallocator.HostSubnetAllocator
-
-	SCTPSupport bool
-
-	// For TCP, UDP, and SCTP type traffic, cache OVN load-balancers used for the
-	// cluster's east-west traffic.
-	loadbalancerClusterCache map[kapi.Protocol]string
-
-	// A cache of all logical switches seen by the watcher and their subnets
-	lsManager *lsm.LogicalSwitchManager
-
-	// A cache of all logical ports known to the controller
-	logicalPortCache *portCache
-
-	// Info about known namespaces. You must use oc.getNamespaceLocked() or
-	// oc.waitForNamespaceLocked() to read this map, and oc.createNamespaceLocked()
-	// or oc.deleteNamespaceLocked() to modify it. namespacesMutex is only held
-	// from inside those functions.
-	namespaces      map[string]*namespaceInfo
-	namespacesMutex sync.Mutex
-
-	externalGWCache map[ktypes.NamespacedName]*externalRouteInfo
-	exGWCacheMutex  sync.RWMutex
-
-	// egressFirewalls is a map of namespaces and the egressFirewall attached to it
-	egressFirewalls sync.Map
-
-	// EgressQoS
-	egressQoSLister egressqoslisters.EgressQoSLister
-	egressQoSSynced cache.InformerSynced
-	egressQoSQueue  workqueue.RateLimitingInterface
-	egressQoSCache  sync.Map
-
-	egressQoSPodLister corev1listers.PodLister
-	egressQoSPodSynced cache.InformerSynced
-	egressQoSPodQueue  workqueue.RateLimitingInterface
-
-	egressQoSNodeLister corev1listers.NodeLister
-	egressQoSNodeSynced cache.InformerSynced
-	egressQoSNodeQueue  workqueue.RateLimitingInterface
-
-	// An address set factory that creates address sets
-	addressSetFactory addressset.AddressSetFactory
-
-	// network policies map, key should be retrieved with getPolicyKey(policy *knet.NetworkPolicy).
-	// network policies that failed to be created will also be added here, and can be retried or cleaned up later.
-	// network policy is only deleted from this map after successful cleanup.
-	// Allowed order of locking is namespace Lock -> oc.networkPolicies key Lock -> networkPolicy.Lock
-	// Don't take namespace Lock while holding networkPolicy key lock to avoid deadlock.
-	networkPolicies *syncmap.SyncMap[*networkPolicy]
-
-	// map of existing shared port groups for network policies
-	// port group exists in the db if and only if port group key is present in this map
-	// key is namespace
-	// allowed locking order is namespace Lock -> networkPolicy.Lock -> sharedNetpolPortGroups key Lock
-	// make sure to keep this order to avoid deadlocks
-	sharedNetpolPortGroups *syncmap.SyncMap[*defaultDenyPortGroups]
-
-	// Supports multicast?
-	multicastSupport bool
-
-	// Cluster wide Load_Balancer_Group UUID.
-	loadBalancerGroupUUID string
-
-	// Cluster-wide router default Control Plane Protection (COPP) UUID
-	defaultCOPPUUID string
-
-	// Controller used for programming OVN for egress IP
-	eIPC egressIPController
-
-	// Controller used to handle services
-	svcController *svccontroller.Controller
-	// Controller used to handle egress services
-	egressSvcController *egresssvc.Controller
-	// svcFactory used to handle service related events
-	svcFactory informers.SharedInformerFactory
-
-	egressFirewallDNS *EgressDNS
-
-	// Is ACL logging enabled while configuring meters?
-	aclLoggingEnabled bool
-
-	joinSwIPManager *lsm.JoinSwitchIPManager
+	podRecorder  *metrics.PodRecorder
 
 	// event recorder used to post events to k8s
 	recorder record.EventRecorder
@@ -162,46 +63,8 @@ type Controller struct {
 	// libovsdb southbound client interface
 	sbClient libovsdbclient.Client
 
-	// retry framework for pods
-	retryPods *retry.RetryFramework
-
-	// retry framework for network policies
-	retryNetworkPolicies *retry.RetryFramework
-
-	// retry framework for egress firewall
-	retryEgressFirewalls *retry.RetryFramework
-
-	// retry framework for egress IP
-	retryEgressIPs *retry.RetryFramework
-	// retry framework for egress IP Namespaces
-	retryEgressIPNamespaces *retry.RetryFramework
-	// retry framework for egress IP Pods
-	retryEgressIPPods *retry.RetryFramework
-	// retry framework for Egress nodes
-	retryEgressNodes *retry.RetryFramework
-	// EgressIP Node-specific syncMap used by egressip node event handler
-	addEgressNodeFailed sync.Map
-
-	// retry framework for nodes
-	retryNodes *retry.RetryFramework
-	// Node-specific syncMaps used by node event handler
-	gatewaysFailed              sync.Map
-	mgmtPortFailed              sync.Map
-	addNodeFailed               sync.Map
-	nodeClusterRouterPortFailed sync.Map
-	hybridOverlayFailed         sync.Map
-
-	// retry framework for Cloud private IP config
-	retryCloudPrivateIPConfig *retry.RetryFramework
-
-	// retry framework for namespaces
-	retryNamespaces *retry.RetryFramework
-
-	podRecorder metrics.PodRecorder
-
-	// variable to determine if all pods present on the node during startup have been processed
-	// updated atomically
-	allInitialPodsProcessed uint32
+	// has SCTP support
+	SCTPSupport bool
 }
 
 const (
@@ -220,239 +83,27 @@ func getPodNamespacedName(pod *kapi.Pod) string {
 	return util.GetLogicalPortName(pod.Namespace, pod.Name)
 }
 
-// NewOvnController creates a new OVN controller for creating logical network
-// infrastructure and policy
-func NewOvnController(ovnClient *util.OVNClientset, wf *factory.WatchFactory, stopChan <-chan struct{}, addressSetFactory addressset.AddressSetFactory,
-	libovsdbOvnNBClient libovsdbclient.Client, libovsdbOvnSBClient libovsdbclient.Client,
-	recorder record.EventRecorder) *Controller {
-	if addressSetFactory == nil {
-		addressSetFactory = addressset.NewOvnAddressSetFactory(libovsdbOvnNBClient)
+// NewBaseNetworkController creates BaseNetworkController shared by controllers
+func NewBaseNetworkController(client clientset.Interface, kube kube.Interface, wf *factory.WatchFactory,
+	recorder record.EventRecorder, nbClient libovsdbclient.Client,
+	sbClient libovsdbclient.Client, podRecorder *metrics.PodRecorder, SCTPSupport bool) *BaseNetworkController {
+	return &BaseNetworkController{
+		client:       client,
+		kube:         kube,
+		watchFactory: wf,
+		recorder:     recorder,
+		nbClient:     nbClient,
+		sbClient:     sbClient,
+		podRecorder:  podRecorder,
+		SCTPSupport:  SCTPSupport,
 	}
-	svcController, svcFactory := newServiceController(ovnClient.KubeClient, libovsdbOvnNBClient, recorder)
-	egressSvcController := newEgressServiceController(ovnClient.KubeClient, libovsdbOvnNBClient, svcFactory, stopChan)
-	var hybridOverlaySubnetAllocator *subnetallocator.HostSubnetAllocator
-	if config.HybridOverlay.Enabled {
-		hybridOverlaySubnetAllocator = subnetallocator.NewHostSubnetAllocator()
-	}
-	oc := &Controller{
-		client: ovnClient.KubeClient,
-		kube: &kube.Kube{
-			KClient:              ovnClient.KubeClient,
-			EIPClient:            ovnClient.EgressIPClient,
-			EgressFirewallClient: ovnClient.EgressFirewallClient,
-			CloudNetworkClient:   ovnClient.CloudNetworkClient,
-		},
-		watchFactory:                 wf,
-		stopChan:                     stopChan,
-		masterSubnetAllocator:        subnetallocator.NewHostSubnetAllocator(),
-		hybridOverlaySubnetAllocator: hybridOverlaySubnetAllocator,
-		lsManager:                    lsm.NewLogicalSwitchManager(),
-		logicalPortCache:             newPortCache(stopChan),
-		namespaces:                   make(map[string]*namespaceInfo),
-		namespacesMutex:              sync.Mutex{},
-		externalGWCache:              make(map[ktypes.NamespacedName]*externalRouteInfo),
-		exGWCacheMutex:               sync.RWMutex{},
-		addressSetFactory:            addressSetFactory,
-		networkPolicies:              syncmap.NewSyncMap[*networkPolicy](),
-		sharedNetpolPortGroups:       syncmap.NewSyncMap[*defaultDenyPortGroups](),
-		eIPC: egressIPController{
-			egressIPAssignmentMutex:           &sync.Mutex{},
-			podAssignmentMutex:                &sync.Mutex{},
-			podAssignment:                     make(map[string]*podAssignmentState),
-			pendingCloudPrivateIPConfigsMutex: &sync.Mutex{},
-			pendingCloudPrivateIPConfigsOps:   make(map[string]map[string]*cloudPrivateIPConfigOp),
-			allocator:                         allocator{&sync.Mutex{}, make(map[string]*egressNode)},
-			nbClient:                          libovsdbOvnNBClient,
-			watchFactory:                      wf,
-			egressIPTotalTimeout:              config.OVNKubernetesFeature.EgressIPReachabiltyTotalTimeout,
-			egressIPNodeHealthCheckPort:       config.OVNKubernetesFeature.EgressIPNodeHealthCheckPort,
-		},
-		loadbalancerClusterCache: make(map[kapi.Protocol]string),
-		multicastSupport:         config.EnableMulticast,
-		loadBalancerGroupUUID:    "",
-		aclLoggingEnabled:        true,
-		joinSwIPManager:          nil,
-		recorder:                 recorder,
-		nbClient:                 libovsdbOvnNBClient,
-		sbClient:                 libovsdbOvnSBClient,
-		svcController:            svcController,
-		svcFactory:               svcFactory,
-		egressSvcController:      egressSvcController,
-		podRecorder:              metrics.NewPodRecorder(),
-	}
-
-	oc.initRetryFrameworkForMaster()
-
-	return oc
-}
-
-func (oc *Controller) initRetryFrameworkForMaster() {
-	// Init the retry framework for pods, namespaces, nodes, network policies, egress firewalls,
-	// egress IP (and dependent namespaces, pods, nodes), cloud private ip config.
-	oc.retryPods = oc.newRetryFrameworkMaster(factory.PodType)
-	oc.retryNetworkPolicies = oc.newRetryFrameworkMaster(factory.PolicyType)
-	oc.retryNodes = oc.newRetryFrameworkMaster(factory.NodeType)
-	oc.retryEgressFirewalls = oc.newRetryFrameworkMaster(factory.EgressFirewallType)
-	oc.retryEgressIPs = oc.newRetryFrameworkMaster(factory.EgressIPType)
-	oc.retryEgressIPNamespaces = oc.newRetryFrameworkMaster(factory.EgressIPNamespaceType)
-	oc.retryEgressIPPods = oc.newRetryFrameworkMaster(factory.EgressIPPodType)
-	oc.retryEgressNodes = oc.newRetryFrameworkMaster(factory.EgressNodeType)
-	oc.retryCloudPrivateIPConfig = oc.newRetryFrameworkMaster(factory.CloudPrivateIPConfigType)
-	oc.retryNamespaces = oc.newRetryFrameworkMaster(factory.NamespaceType)
-}
-
-// Run starts the actual watching.
-func (oc *Controller) Run(ctx context.Context, wg *sync.WaitGroup) error {
-	// Start and sync the watch factory to begin listening for events
-	if err := oc.watchFactory.Start(); err != nil {
-		return err
-	}
-
-	oc.syncPeriodic()
-	klog.Infof("Starting all the Watchers...")
-	start := time.Now()
-
-	// Sync external gateway routes. External gateway may be set in namespaces
-	// or via pods. So execute an individual sync method at startup
-	oc.cleanExGwECMPRoutes()
-
-	// WatchNamespaces() should be started first because it has no other
-	// dependencies, and WatchNodes() depends on it
-	if err := oc.WatchNamespaces(); err != nil {
-		return err
-	}
-
-	// WatchNodes must be started next because it creates the node switch
-	// which most other watches depend on.
-	// https://github.com/ovn-org/ovn-kubernetes/pull/859
-	if err := oc.WatchNodes(); err != nil {
-		return err
-	}
-
-	// Start service watch factory and sync services
-	oc.svcFactory.Start(oc.stopChan)
-
-	// Services should be started after nodes to prevent LB churn
-	if err := oc.StartServiceController(wg, true); err != nil {
-		return err
-	}
-
-	if err := oc.WatchPods(); err != nil {
-		return err
-	}
-
-	// WatchNetworkPolicy depends on WatchPods and WatchNamespaces
-	if err := oc.WatchNetworkPolicy(); err != nil {
-		return err
-	}
-
-	if config.OVNKubernetesFeature.EnableEgressIP {
-		// This is probably the best starting order for all egress IP handlers.
-		// WatchEgressIPNamespaces and WatchEgressIPPods only use the informer
-		// cache to retrieve the egress IPs when determining if namespace/pods
-		// match. It is thus better if we initialize them first and allow
-		// WatchEgressNodes / WatchEgressIP to initialize after. Those handlers
-		// might change the assignments of the existing objects. If we do the
-		// inverse and start WatchEgressIPNamespaces / WatchEgressIPPod last, we
-		// risk performing a bunch of modifications on the EgressIP objects when
-		// we restart and then have these handlers act on stale data when they
-		// sync.
-		if err := oc.WatchEgressIPNamespaces(); err != nil {
-			return err
-		}
-		if err := oc.WatchEgressIPPods(); err != nil {
-			return err
-		}
-		if err := oc.WatchEgressNodes(); err != nil {
-			return err
-		}
-		if err := oc.WatchEgressIP(); err != nil {
-			return err
-		}
-		if util.PlatformTypeIsEgressIPCloudProvider() {
-			if err := oc.WatchCloudPrivateIPConfig(); err != nil {
-				return err
-			}
-		}
-		if config.OVNKubernetesFeature.EgressIPReachabiltyTotalTimeout == 0 {
-			klog.V(2).Infof("EgressIP node reachability check disabled")
-		} else if config.OVNKubernetesFeature.EgressIPNodeHealthCheckPort != 0 {
-			klog.Infof("EgressIP node reachability enabled and using gRPC port %d",
-				config.OVNKubernetesFeature.EgressIPNodeHealthCheckPort)
-		}
-	}
-
-	if config.OVNKubernetesFeature.EnableEgressFirewall {
-		var err error
-		oc.egressFirewallDNS, err = NewEgressDNS(oc.addressSetFactory, oc.stopChan)
-		if err != nil {
-			return err
-		}
-		oc.egressFirewallDNS.Run(egressFirewallDNSDefaultDuration)
-		err = oc.WatchEgressFirewall()
-		if err != nil {
-			return err
-		}
-	}
-
-	if config.OVNKubernetesFeature.EnableEgressQoS {
-		oc.initEgressQoSController(
-			oc.watchFactory.EgressQoSInformer(),
-			oc.watchFactory.PodCoreInformer(),
-			oc.watchFactory.NodeCoreInformer())
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			oc.runEgressQoSController(1, oc.stopChan)
-		}()
-	}
-
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		oc.egressSvcController.Run(1)
-	}()
-
-	klog.Infof("Completing all the Watchers took %v", time.Since(start))
-
-	if config.Kubernetes.OVNEmptyLbEvents {
-		klog.Infof("Starting unidling controller")
-		unidlingController, err := unidling.NewController(
-			oc.recorder,
-			oc.watchFactory.ServiceInformer(),
-			oc.sbClient,
-		)
-		if err != nil {
-			return err
-		}
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			unidlingController.Run(oc.stopChan)
-		}()
-	}
-
-	// Final step to cleanup after resource handlers have synced
-	err := oc.ovnTopologyCleanup()
-	if err != nil {
-		klog.Errorf("Failed to cleanup OVN topology to version %d: %v", ovntypes.OvnCurrentTopologyVersion, err)
-		return err
-	}
-
-	// Master is fully running and resource handlers have synced, update Topology version in OVN and the ConfigMap
-	if err := oc.reportTopologyVersion(ctx); err != nil {
-		klog.Errorf("Failed to report topology version: %v", err)
-		return err
-	}
-
-	return nil
 }
 
 // syncPeriodic adds a goroutine that periodically does some work
 // right now there is only one ticker registered
 // for syncNodesPeriodic which deletes chassis records from the sbdb
 // every 5 minutes
-func (oc *Controller) syncPeriodic() {
+func (oc *DefaultNetworkController) syncPeriodic() {
 	go func() {
 		nodeSyncTicker := time.NewTicker(5 * time.Minute)
 		defer nodeSyncTicker.Stop()
@@ -467,14 +118,33 @@ func (oc *Controller) syncPeriodic() {
 	}()
 }
 
-func (oc *Controller) recordPodEvent(addErr error, pod *kapi.Pod) {
+func (oc *DefaultNetworkController) getPortInfo(pod *kapi.Pod) *lpInfo {
+	var portInfo *lpInfo
+	key := util.GetLogicalPortName(pod.Namespace, pod.Name)
+	if !util.PodWantsNetwork(pod) {
+		// create dummy logicalPortInfo for host-networked pods
+		mac, _ := net.ParseMAC("00:00:00:00:00:00")
+		portInfo = &lpInfo{
+			logicalSwitch: "host-networked",
+			name:          key,
+			uuid:          "host-networked",
+			ips:           []*net.IPNet{},
+			mac:           mac,
+		}
+	} else {
+		portInfo, _ = oc.logicalPortCache.get(key)
+	}
+	return portInfo
+}
+
+func (oc *DefaultNetworkController) recordPodEvent(reason string, addErr error, pod *kapi.Pod) {
 	podRef, err := ref.GetReference(scheme.Scheme, pod)
 	if err != nil {
 		klog.Errorf("Couldn't get a reference to pod %s/%s to post an event: '%v'",
 			pod.Namespace, pod.Name, err)
 	} else {
 		klog.V(5).Infof("Posting a %s event for Pod %s/%s", kapi.EventTypeWarning, pod.Namespace, pod.Name)
-		oc.recorder.Eventf(podRef, kapi.EventTypeWarning, "ErrorAddingLogicalPort", addErr.Error())
+		oc.recorder.Eventf(podRef, kapi.EventTypeWarning, reason, addErr.Error())
 	}
 }
 
@@ -490,7 +160,7 @@ func networkStatusAnnotationsChanged(oldPod, newPod *kapi.Pod) bool {
 
 // ensurePod tries to set up a pod. It returns nil on success and error on failure; failure
 // indicates the pod set up should be retried later.
-func (oc *Controller) ensurePod(oldPod, pod *kapi.Pod, addPort bool) error {
+func (oc *DefaultNetworkController) ensurePod(oldPod, pod *kapi.Pod, addPort bool) error {
 	// Try unscheduled pods later
 	if !util.PodScheduled(pod) {
 		return nil
@@ -523,7 +193,7 @@ func (oc *Controller) ensurePod(oldPod, pod *kapi.Pod, addPort bool) error {
 
 // removePod tried to tear down a pod. It returns nil on success and error on failure;
 // failure indicates the pod tear down should be retried later.
-func (oc *Controller) removePod(pod *kapi.Pod, portInfo *lpInfo) error {
+func (oc *DefaultNetworkController) removePod(pod *kapi.Pod, portInfo *lpInfo) error {
 	if !util.PodWantsNetwork(pod) {
 		if err := oc.deletePodExternalGW(pod); err != nil {
 			return fmt.Errorf("unable to delete external gateway routes for pod %s: %w",
@@ -539,35 +209,35 @@ func (oc *Controller) removePod(pod *kapi.Pod, portInfo *lpInfo) error {
 }
 
 // WatchPods starts the watching of the Pod resource and calls back the appropriate handler logic
-func (oc *Controller) WatchPods() error {
+func (oc *DefaultNetworkController) WatchPods() error {
 	_, err := oc.retryPods.WatchResource()
 	return err
 }
 
 // WatchNetworkPolicy starts the watching of the network policy resource and calls
 // back the appropriate handler logic
-func (oc *Controller) WatchNetworkPolicy() error {
+func (oc *DefaultNetworkController) WatchNetworkPolicy() error {
 	_, err := oc.retryNetworkPolicies.WatchResource()
 	return err
 }
 
 // WatchEgressFirewall starts the watching of egressfirewall resource and calls
 // back the appropriate handler logic
-func (oc *Controller) WatchEgressFirewall() error {
+func (oc *DefaultNetworkController) WatchEgressFirewall() error {
 	_, err := oc.retryEgressFirewalls.WatchResource()
 	return err
 }
 
 // WatchEgressNodes starts the watching of egress assignable nodes and calls
 // back the appropriate handler logic.
-func (oc *Controller) WatchEgressNodes() error {
+func (oc *DefaultNetworkController) WatchEgressNodes() error {
 	_, err := oc.retryEgressNodes.WatchResource()
 	return err
 }
 
 // WatchCloudPrivateIPConfig starts the watching of cloudprivateipconfigs
 // resource and calls back the appropriate handler logic.
-func (oc *Controller) WatchCloudPrivateIPConfig() error {
+func (oc *DefaultNetworkController) WatchCloudPrivateIPConfig() error {
 	_, err := oc.retryCloudPrivateIPConfig.WatchResource()
 	return err
 }
@@ -575,30 +245,30 @@ func (oc *Controller) WatchCloudPrivateIPConfig() error {
 // WatchEgressIP starts the watching of egressip resource and calls back the
 // appropriate handler logic. It also initiates the other dedicated resource
 // handlers for egress IP setup: namespaces, pods.
-func (oc *Controller) WatchEgressIP() error {
+func (oc *DefaultNetworkController) WatchEgressIP() error {
 	_, err := oc.retryEgressIPs.WatchResource()
 	return err
 }
 
-func (oc *Controller) WatchEgressIPNamespaces() error {
+func (oc *DefaultNetworkController) WatchEgressIPNamespaces() error {
 	_, err := oc.retryEgressIPNamespaces.WatchResource()
 	return err
 }
 
-func (oc *Controller) WatchEgressIPPods() error {
+func (oc *DefaultNetworkController) WatchEgressIPPods() error {
 	_, err := oc.retryEgressIPPods.WatchResource()
 	return err
 }
 
 // WatchNamespaces starts the watching of namespace resource and calls
 // back the appropriate handler logic
-func (oc *Controller) WatchNamespaces() error {
+func (oc *DefaultNetworkController) WatchNamespaces() error {
 	_, err := oc.retryNamespaces.WatchResource()
 	return err
 }
 
 // syncNodeGateway ensures a node's gateway router is configured
-func (oc *Controller) syncNodeGateway(node *kapi.Node, hostSubnets []*net.IPNet) error {
+func (oc *DefaultNetworkController) syncNodeGateway(node *kapi.Node, hostSubnets []*net.IPNet) error {
 	l3GatewayConfig, err := util.ParseNodeL3GatewayAnnotation(node)
 	if err != nil {
 		return err
@@ -635,7 +305,7 @@ func (oc *Controller) syncNodeGateway(node *kapi.Node, hostSubnets []*net.IPNet)
 
 // WatchNodes starts the watching of node resource and calls
 // back the appropriate handler logic
-func (oc *Controller) WatchNodes() error {
+func (oc *DefaultNetworkController) WatchNodes() error {
 	_, err := oc.retryNodes.WatchResource()
 	return err
 }
@@ -657,7 +327,7 @@ func (oc *Controller) WatchNodes() error {
 // *) If one of "allow" or "deny" can be parsed and has a valid value, but the other key is not present in the
 //
 //	annotation, then assume that this key should be disabled by setting its nsInfo value to "".
-func (oc *Controller) aclLoggingUpdateNsInfo(annotation string, nsInfo *namespaceInfo) error {
+func (oc *DefaultNetworkController) aclLoggingUpdateNsInfo(annotation string, nsInfo *namespaceInfo) error {
 	var aclLevels ACLLoggingLevels
 	var errors []error
 
@@ -793,7 +463,7 @@ func newServiceController(client clientset.Interface, nbClient libovsdbclient.Cl
 	return controller, svcFactory
 }
 
-func (oc *Controller) StartServiceController(wg *sync.WaitGroup, runRepair bool) error {
+func (oc *DefaultNetworkController) StartServiceController(wg *sync.WaitGroup, runRepair bool) error {
 	klog.Infof("Starting OVN Service Controller: Using Endpoint Slices")
 	wg.Add(1)
 	go func() {
