+++ go-controller/pkg/retry/obj_retry.go
@@ -56,7 +56,7 @@ type EventHandler interface {
 	RecordUpdateEvent(obj interface{})
 	RecordDeleteEvent(obj interface{})
 	RecordSuccessEvent(obj interface{})
-	RecordErrorEvent(obj interface{}, err error)
+	RecordErrorEvent(obj interface{}, reason string, err error)
 }
 
 type ResourceHandler struct {
@@ -285,7 +285,7 @@ func (r *RetryFramework) resourceRetry(objKey string, now time.Time) {
 		} else {
 			// delete old object if needed
 			if entry.oldObj != nil {
-				klog.Infof("Removing old object: %s %s (failed: %s)",
+				klog.Infof("Removing old object: %s %s (failed: %v)",
 					r.ResourceHandler.ObjType, objKey, entry.failedAttempts)
 				if !r.ResourceHandler.IsResourceScheduled(entry.oldObj) {
 					klog.V(5).Infof("Retry: %s %s not scheduled", r.ResourceHandler.ObjType, objKey)
@@ -408,7 +408,7 @@ func (r *RetryFramework) processObjectInTerminalState(obj interface{}, lockedKey
 	if err := r.ResourceHandler.DeleteResource(obj, internalCacheEntry); err != nil {
 		klog.Errorf("Failed to delete object %s of type %s in terminal state, during %s event: %v",
 			lockedKey, r.ResourceHandler.ObjType, event, err)
-		r.ResourceHandler.RecordErrorEvent(obj, err)
+		r.ResourceHandler.RecordErrorEvent(obj, "ErrorDeletingResource", err)
 		r.increaseFailedAttemptsCounter(retryEntry)
 		return
 	}
@@ -466,7 +466,7 @@ func (r *RetryFramework) WatchResourceFiltered(namespaceForFilteredHandler strin
 						if err := r.ResourceHandler.DeleteResource(retryObj.oldObj, internalCacheEntry); err != nil {
 							klog.Errorf("Failed to delete old object %s of type %s,"+
 								" during add event: %v", key, r.ResourceHandler.ObjType, err)
-							r.ResourceHandler.RecordErrorEvent(obj, err)
+							r.ResourceHandler.RecordErrorEvent(obj, "ErrorDeletingResource", err)
 							r.increaseFailedAttemptsCounter(retryObj)
 							return
 						}
@@ -475,7 +475,7 @@ func (r *RetryFramework) WatchResourceFiltered(namespaceForFilteredHandler strin
 					start := time.Now()
 					if err := r.ResourceHandler.AddResource(obj, false); err != nil {
 						klog.Errorf("Failed to create %s %s, error: %v", r.ResourceHandler.ObjType, key, err)
-						r.ResourceHandler.RecordErrorEvent(obj, err)
+						r.ResourceHandler.RecordErrorEvent(obj, "ErrorAddingResource", err)
 						r.increaseFailedAttemptsCounter(retryObj)
 						return
 					}
@@ -557,7 +557,7 @@ func (r *RetryFramework) WatchResourceFiltered(namespaceForFilteredHandler strin
 						if err := r.ResourceHandler.DeleteResource(retryEntryOrNil.oldObj,
 							retryEntryOrNil.config); err != nil {
 							klog.Errorf("Failed to delete stale object %s, during update: %v", oldKey, err)
-							r.ResourceHandler.RecordErrorEvent(retryEntryOrNil.oldObj, err)
+							r.ResourceHandler.RecordErrorEvent(retryEntryOrNil.oldObj, "ErrorDeletingResource", err)
 							retryEntry := r.initRetryObjWithAdd(latest, key)
 							r.increaseFailedAttemptsCounter(retryEntry)
 							return
@@ -584,7 +584,7 @@ func (r *RetryFramework) WatchResourceFiltered(namespaceForFilteredHandler strin
 						if err := r.ResourceHandler.DeleteResource(old, existingCacheEntry); err != nil {
 							klog.Errorf("Failed to delete %s %s, during update: %v",
 								r.ResourceHandler.ObjType, oldKey, err)
-							r.ResourceHandler.RecordErrorEvent(old, err)
+							r.ResourceHandler.RecordErrorEvent(old, "ErrorDeletingResource", err)
 							retryEntry := r.InitRetryObjWithDelete(old, key, nil, false)
 							r.initRetryObjWithAdd(latest, key)
 							r.increaseFailedAttemptsCounter(retryEntry)
@@ -603,7 +603,7 @@ func (r *RetryFramework) WatchResourceFiltered(namespaceForFilteredHandler strin
 						if err := r.ResourceHandler.UpdateResource(old, latest, found); err != nil {
 							klog.Errorf("Failed to update %s, old=%s, new=%s, error: %v",
 								r.ResourceHandler.ObjType, oldKey, newKey, err)
-							r.ResourceHandler.RecordErrorEvent(latest, err)
+							r.ResourceHandler.RecordErrorEvent(latest, "ErrorUpdatingResource", err)
 							var retryEntry *retryObjEntry
 							if r.ResourceHandler.NeedsUpdateDuringRetry {
 								retryEntry = r.initRetryObjWithUpdate(old, latest, key)
@@ -615,7 +615,7 @@ func (r *RetryFramework) WatchResourceFiltered(namespaceForFilteredHandler strin
 						}
 					} else { // we previously deleted old object, now let's add the new one
 						if err := r.ResourceHandler.AddResource(latest, false); err != nil {
-							r.ResourceHandler.RecordErrorEvent(latest, err)
+							r.ResourceHandler.RecordErrorEvent(latest, "ErrorAddingResource", err)
 							retryEntry := r.initRetryObjWithAdd(latest, key)
 							r.increaseFailedAttemptsCounter(retryEntry)
 							klog.Errorf("Failed to add %s %s, during update: %v",
